ImagePager -> ImageCarousel로 변경하고 UIModel 만든 후 SimpleBindingAdapter, SimpleBindingViewHolder 사용
