+++ presentation/src/main/java/com/dev/playground/presentation/base/SimpleBindingAdapter.kt
@@ -2,9 +2,11 @@ package com.dev.playground.presentation.base
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import android.widget.ImageView
 import androidx.databinding.DataBindingUtil
 import androidx.databinding.ViewDataBinding
 import androidx.recyclerview.widget.ListAdapter
+import com.bumptech.glide.Glide
 import com.dev.playground.presentation.model.base.SimpleItemDiffCallback
 import com.dev.playground.presentation.model.base.SimpleUIModel
 
@@ -28,6 +30,13 @@ open class SimpleBindingAdapter<T, VH>(
 
     override fun onBindViewHolder(holder: VH, position: Int) {
         holder.onBind(getItem(position))
+    }
+
+    override fun onViewRecycled(holder: VH) {
+        super.onViewRecycled(holder)
+        (holder.itemView as? ImageView)?.run {
+            Glide.with(this).clear(this)
+        }
     }
 
     override fun getItemViewType(position: Int): Int = getItem(position).layoutRes
