+++ presentation/src/main/java/com/dev/playground/presentation/model/ImageCarouselItemUIModel.kt
@@ -1,0 +1,6 @@
+package com.dev.playground.presentation.model
+
+import com.dev.playground.presentation.R
+import com.dev.playground.presentation.model.base.SimpleUIModel
+
+data class ImageCarouselItemUIModel(val url: String): SimpleUIModel(R.layout.item_image_carousel)
