+++ presentation/src/main/res/layout/view_image_carousel.xml
